% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minmaxPareto.R
\name{minmaxPareto}
\alias{minmaxPareto}
\title{Pareto maximum and minimum values}
\usage{
minmaxPareto(osc, ...)
}
\arguments{
\item{osc}{A list with the optimized sample configurations (OSC). Each OSC
must be named after the objective function with which it has been optimized.
For example, \code{osc = list(CORR = osc_corr, DIST = osc_dist)}.}

\item{...}{Other arguments required to compute the objective function value.
See the help pages of the respective objective functions to see which
arguments are needed.}
}
\value{
A data frame with the Pareto maximum and minimum values.
}
\description{
Compute the Pareto maximum and minimum values of the objective functions
that compose a multi-objective optimization problem (MOOP). The Pareto
maximum and minimum values are then used to transform the objective
functions using the \emph{upper-lower-bound approach} so that they can be
aggregated into a single \emph{utility function} using the \emph{weighted
sum method}.
}
\examples{
require(sp)
data(meuse.grid)
candi <- meuse.grid[, 1:2]
covars <- meuse.grid[, c(1, 2, 5)]
schedule <- scheduleSPSANN(initial.temperature = 1, chains = 1)
set.seed(2001)
osc_corr <- optimCORR(points = 10, candi = candi, covars = covars,
                      schedule = schedule)
set.seed(2001)
osc_dist <- optimDIST(points = 10, candi = candi, covars = covars,
                      schedule = schedule)
pareto <- minmaxPareto(osc = list(DIST = osc_dist, CORR = osc_corr),
                       candi = candi, covars = covars)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\seealso{
\code{\link[spsann]{optimACDC}}, \code{\link[spsann]{SPAN}}
}

