% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimACDC.R
\name{optimACDC}
\alias{optimACDC}
\title{Optimization of sample patterns for trend estimation}
\usage{
optimACDC(points, candi, covars, strata.type = "area", weights = list(correl
  = 0.5, strata = 0.5), use.coords = FALSE, nadir = list(sim = 1000,
  save.sim = TRUE, user = NULL, abs = NULL), utopia = list(user = list(correl
  = NULL, strata = NULL), abs = NULL), scale = list(type = "upper-lower", max
  = 100), x.max, x.min, y.max, y.min, iterations, acceptance = list(initial =
  0.99, cooling = iterations/10), stopping = list(max.count = iterations/10),
  plotit = TRUE, boundary, progress = TRUE, verbose = TRUE)
}
\arguments{
\item{points}{Integer value, integer vector, data frame or matrix. If
\code{points} is an integer value, it defines the number of points that
should be sampled from \code{candi}. If \code{points} is an integer vector,
it contains the row indexes of \code{candi} that correspond to the points.
If \code{points} is a data frame or matrix, it must have three
columns: \code{[, 1]} the identification of each sample point (1, 2, ..., n),
\code{[, 2]} the x-coordinates, and \code{[, 3]} the y-coordinates. The
coordinates must be projected. If a data frame or matrix is used,
\code{points} must be a subset of \code{candi}.}

\item{candi}{Data frame or matrix with the candidate locations for the
sample points. The coordinates must be projected.}

\item{covars}{Data frame or matrix with the covariates in the columns.}

\item{strata.type}{Character value. The type of strata to be used with
numeric covariates. Available options are \code{"area"} for equal area and
\code{"range"} for equal range. Defaults to \code{strata.type = "area"}. See
\sQuote{Details} for more information.}

\item{weights}{List with two named sub-arguments. The weights assigned to
correlation/association measure and the sampling strata/classes. They
must be named and sum to unity. Defaults to
\code{weights = list(correl = 0.5, strata = 0.5)}.}

\item{use.coords}{Logical value. Should the coordinates be used as
covariates? Defaults to \code{use.coords = FALSE}.}

\item{nadir}{List with four named sub-arguments: \code{sim} -- the number of
random realizations to estimate the nadir point; \code{save.sim} -- logical
for saving the simulated values and returning them as an attribute of the
optimized sample pattern; \code{user} -- a user-defined value;
\code{abs} -- logical for calculating the nadir point internally. Only
simulations are implemented in the current version. Defaults to
\code{nadir = list(sim = 1000, save.sim = TRUE, user = NULL, abs = NULL)}.}

\item{utopia}{List with two named sub-arguments: \code{user} -- a list of
two user-defined values (\code{correl} and \code{strata}), and \code{abs} --
a logical value for calculating the utopia point internally. Defaults to
\code{utopia = list(user = list(correl = NULL, strata = NULL), abs = NULL)}.}

\item{scale}{List with two named sub-arguments: \code{type} -- the type of
scaling that should be used, with available options \code{"upper} and
\code{"upper-lower}; and \code{max} -- the maximum value after scaling.
Defaults to \code{scale = list(type = "upper-lower", max = 100)}.}

\item{x.max,x.min,y.max,y.min}{Numeric value. The minimum and maximum
quantity of random noise to be added to the x and y coordinates. The minimum
quantity must be equal to, at least, the minimum distance between two
neighbouring candidate locations. The units are the same as of the
coordinates.}

\item{iterations}{Integer. The maximum number of iterations that should be
used for the optimization.}

\item{acceptance}{List with two named sub-arguments: \code{initial} --
numeric value between 0 and 1 defining the initial acceptance probability,
and \code{cooling} -- a numeric value defining the exponential factor by
witch the acceptance probability decreases at each iteration. Defaults to
\code{acceptance = list(initial = 0.99, cooling = iterations / 10)}.}

\item{stopping}{List with one named sub-argument: \code{max.count} --
integer value defining the maximum allowable number of iterations without
improvement of the objective function value. Defaults to
\code{stopping = list(max.count = iterations / 10)}. More options may be
included in the future.}

\item{plotit}{Logical for plotting the optimization results. This includes
a) the progress of the objective function values and acceptance
probabilities, and b) the original points, the perturbed points and the
progress of the maximum perturbation in the x- and y-coordinates. The plots
are updated at each 10 iterations. Defaults to \code{plotit = TRUE}.}

\item{boundary}{SpatialPolygon. The boundary of the spatial domain.
Mandatory if \code{plotit = TRUE}.}

\item{progress}{Logical for printing a progress bar. Defaults to
\code{progress = TRUE}.}

\item{verbose}{Logical for printing messages about the progress of the
optimization.}

\item{greedy}{Logical value. Should the optimization be done using a greedy
algorithm, that is, accepting only better system configurations? Defaults
to \code{greedy = FALSE}.}
}
\value{
\code{optimACDC} returns a matrix: the optimized sample pattern with
the evolution of the energy state during the optimization as an attribute.

\code{objACDC} returns a numeric value: the energy state of the point
pattern.
}
\description{
Optimize a sample pattern for trend estimation. The criterion is defined so
that the sample reproduces the association/correlation between the
covariates, as well as their marginal distribution (\bold{ACDC}).
}
\details{
This method is also known as the conditioned Latin Hypercube of Minasny and
McBratney (2006). Visit the package manual to see the corrections that we
have made in that method.
}
\examples{
require(pedometrics)
require(sp)
require(rgeos)
data(meuse.grid)
candi <- meuse.grid[, 1:2]
coordinates(candi) <- ~ x + y
gridded(candi) <- TRUE
boundary <- as(candi, "SpatialPolygons")
boundary <- gUnionCascaded(boundary)
candi <- coordinates(candi)
candi <- matrix(cbind(1:dim(candi)[1], candi), ncol = 3)
covars <- meuse.grid[, 5]
x.max <- diff(bbox(boundary)[1, ])
y.min <- 40
x.min <- 40
y.max <- diff(bbox(boundary)[2, ])
nadir <- list(sim = 10, save.sim = TRUE, user = NULL, abs = NULL)
utopia <- list(user = list(correl = 0, strata = 0), abs = NULL)
scale <- list(type = "upper-lower", max = 100)
weights <- list(strata = 0.5, correl = 0.5)
set.seed(2001)
res <- optimACDC(points = 100, candi = candi, covars = covars,
                 use.coords = TRUE,
                 weights = weights, x.max = x.max, x.min = x.min,
                 y.max = y.max, y.min = y.min, boundary = boundary,
                 nadir = nadir, iterations = 500, utopia = utopia,
                 scale = scale)
tail(attr(res, "energy"))
objACDC(points = res, candi = candi, covars = covars, use.coords = TRUE,
        weights = weights, nadir = nadir,
        utopia = utopia, scale = scale)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\references{
Minasny, B.; McBratney, A. B. A conditioned Latin hypercube method for
sampling in the presence of ancillary information. \emph{Computers &
Geosciences}, v. 32, p. 1378-1388, 2006.

Minasny, B.; McBratney, A. B. Conditioned Latin Hypercube Sampling for
calibrating soil sensor data to soil properties. Chapter 9. Viscarra Rossel,
R. A.; McBratney, A. B.; Minasny, B. (Eds.) \emph{Proximal Soil Sensing}.
Amsterdam: Springer, p. 111-119, 2010.

Mulder, V. L.; de Bruin, S.; Schaepman, M. E. Representing major soil
variability at regional scale by constrained Latin hypercube sampling of
remote sensing data. \emph{International Journal of Applied Earth Observation
and Geoinformation}, v. 21, p. 301-310, 2013.

Roudier, P.; Beaudette, D.; Hewitt, A. A conditioned Latin hypercube sampling
algorithm incorporating operational constraints. \emph{5th Global Workshop on
Digital Soil Mapping}. Sydney, p. 227-231, 2012.
}
\seealso{
\code{\link[clhs]{clhs}}
}
\concept{
simulated annealing
}
\keyword{optimize}
\keyword{spatial}

