% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optimCORR.R
\name{optimCORR}
\alias{objCORR}
\alias{optimCORR}
\title{Optimization of sample patterns for trend estimation}
\usage{
optimCORR(points, candi, covars, covars.type = "numeric",
  use.coords = FALSE, strata.type = "area", x.max, x.min, y.max, y.min,
  iterations, acceptance = list(initial = 0.99, cooling = iterations/10),
  stopping = list(max.count = iterations/10), plotit = TRUE, boundary,
  progress = TRUE, verbose = TRUE)

objCORR(points, candi, covars, covars.type = "numeric", use.coords = FALSE,
  strata.type = "area")
}
\arguments{
\item{points}{Integer value, integer vector, data frame or matrix. If
\code{points} is an integer value, it defines the number of points that
should be sampled from \code{candi}. If \code{points} is an integer vector,
it contains the row indexes of \code{candi} that correspond to the points.
If \code{points} is a data frame or matrix, it must have three
columns: \code{[, 1]} the identification of each point (1, 2, ..., n),
\code{[, 2]} the x coordinates, and \code{[, 3]} the y coordinates. The
coordinates must be projected. If a data frame or matrix is used,
\code{points} must be a subset of \code{candi}.}

\item{candi}{Data frame or matrix with the candidate locations for the
sample points. The coordinates must be projected.}

\item{covars}{Data frame or matrix with the covariates in the columns.}

\item{covars.type}{Character value. The type of covariates that is
being used. Available options are \code{"numeric"} and \code{"factor"}.
Defaults to \code{covars.type = "numeric"}.}

\item{use.coords}{Logical value. Should the coordinates be used as
covariates? Defaults to \code{use.coords = FALSE}.}

\item{strata.type}{Character value. The type of strata to be used to
categorize the coordinates when they are used with covariates of type factor.
Available options are \code{"area"} for equal area and \code{"range"} for
equal range. Defaults to \code{strata.type = "area"}.}

\item{x.max,x.min,y.max,y.min}{Numeric value. The minimum and maximum
quantity of random noise to be added to the x and y coordinates. The minimum
quantity must be equal to, at least, the minimum distance between two
neighboring candidate locations. The units are the same as of the coordinates.}

\item{iterations}{Integer. The maximum number of iterations that should be
used for the optimization.}

\item{acceptance}{List with two named sub-arguments: \code{initial} --
numeric value between 0 and 1 defining the initial acceptance probability,
and \code{cooling} -- a numeric value defining the exponential factor by
witch the acceptance probability decreases at each iteration. Defaults to
\code{acceptance = list(initial = 0.99, cooling = iterations / 10)}.}

\item{stopping}{List with one named sub-argument: \code{max.count} --
integer value defining the maximum allowable number of iterations without
improvement of the objective function value. Defaults to
\code{stopping = list(max.count = iterations / 10)}.}

\item{plotit}{Logical for ploting the optimization results. This includes
a) the progress of the objective function values and acceptance
probabilities, and b) the original points, the perturbed points and the
progress of the maximum perturbation in the x and y coordinates. The plots
are updated at each 10 iterations. Defaults to \code{plotit = TRUE}.}

\item{boundary}{SpatialPolygon. The boundary of the spatial domain.
Mandatory if \code{plotit = TRUE}.}

\item{progress}{Logical for printing a progress bar. Defaults to
\code{progress = TRUE}.}

\item{verbose}{Logical for printing messages about the progress of the
optimization.}
}
\value{
\code{optimCORR} returns a matrix: the optimized sample pattern with
the evolution of the energy state during the optimization as an attribute.

\code{objCORR} returns a numeric value: the energy state of the point
pattern.
}
\description{
Optimize a sample pattern for trend estimation. The criterion is defined so
that the sample reproduces the association/correlation between the covariates
(\bold{CORR}).
}
\details{
This method was derived from the conditioned Latin Hypercube of Minasny and
McBratney (2006). Visit the package manual to see the corrections that we
have made in that method.
}
\examples{
require(pedometrics)
require(sp)
require(rgeos)
data(meuse.grid)
candi              <- meuse.grid[, 1:2]
coordinates(candi) <- ~ x + y
gridded(candi)     <- TRUE
boundary           <- as(candi, "SpatialPolygons")
boundary           <- gUnionCascaded(boundary)
candi              <- coordinates(candi)
candi              <- matrix(cbind(1:dim(candi)[1], candi), ncol = 3)
covars             <- meuse.grid[, 5]
x.max              <- diff(bbox(boundary)[1, ])
y.max              <- diff(bbox(boundary)[2, ])
y.min              <- 40
x.min              <- 40
set.seed(2001)
res <- optimCORR(points = 100, candi = candi, covars = covars,
                 use.coords = TRUE, covars.type = "numeric", x.max = x.max,
                 x.min = x.min, y.max = y.max, y.min = y.min,
                 boundary = boundary, iterations = 500)
tail(attr(res, "energy"))
objCORR(points = res, candi = candi, covars = covars,
        covars.type = "numeric", use.coords = TRUE)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\references{
Minasny, B.; McBratney, A. B. A conditioned Latin hypercube method for
sampling in the presence of ancillary information. \emph{Computers &
Geosciences}, v. 32, p. 1378-1388, 2006.

Minasny, B.; McBratney, A. B. Conditioned Latin Hypercube Sampling for
calibrating soil sensor data to soil properties. Chapter 9. Viscarra Rossel,
R. A.; McBratney, A. B.; Minasny, B. (Eds.) \emph{Proximal Soil Sensing}.
Amsterdam: Springer, p. 111-119, 2010.

Mulder, V. L.; de Bruin, S.; Schaepman, M. E. Representing major soil
variability at regional scale by constrained Latin hypercube sampling of
remote sensing data. \emph{International Journal of Applied Earth Observation
and Geoinformation}, v. 21, p. 301-310, 2013.

Roudier, P.; Beaudette, D.; Hewitt, A. A conditioned Latin hypercube sampling
algorithm incorporating operational constraints. \emph{5th Global Workshop on
Digital Soil Mapping}. Sydney, p. 227-231, 2012.
}
\seealso{
\code{\link[clhs]{clhs}}
}
\concept{
simulated annealing
}
\keyword{optimize}
\keyword{spatial}

