% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimDIST.R
\name{optimDIST}
\alias{DIST}
\alias{objDIST}
\alias{optimDIST}
\title{Optimization of sample configurations for spatial trend identification and estimation (II)}
\usage{
optimDIST(points, candi, covars, strata.type = "area", use.coords = FALSE,
  schedule = scheduleSPSANN(), plotit = FALSE, track = FALSE, boundary,
  progress = "txt", verbose = FALSE)

objDIST(points, candi, covars, strata.type = "area", use.coords = FALSE)
}
\arguments{
\item{points}{Integer value, integer vector, data frame or matrix. If \code{points} is an integer value, 
it defines the number of points that should be randomly sampled from \code{candi} to form the starting 
system configuration. If \code{points} is a vector of integer values, it contains the row indexes of 
\code{candi} that correspond to the points that form the starting system configuration. If \code{points} 
is a data frame or matrix, it must have three columns in the following order: \code{[, "id"]} the row 
indexes of \code{candi} that correspond to each point, \code{[, "x"]} the projected x-coordinates, and 
\code{[, "y"]} the projected y-coordinates. Note that in the later case, \code{points} must be a subset 
of \code{candi}.}

\item{candi}{Data frame or matrix with the candidate locations for the jittered points. \code{candi} must 
have two columns in the following order: \code{[, "x"]} the projected x-coordinates, and \code{[, "y"]} the 
projected y-coordinates.}

\item{covars}{Data frame or matrix with the covariates in the columns.}

\item{strata.type}{Character value setting the type of stratification that should be used to create the 
marginal sampling strata (or factor levels) for the numeric covariates. Available options are \code{"area"},
for equal-area, and \code{"range"}, for equal-range. Defaults to \code{strata.type = "area"}.}

\item{use.coords}{Logical value. Should the geographic coordinates be used as covariates? Defaults to 
\code{use.coords = FALSE}.}

\item{schedule}{List with 11 named sub-arguments defining the control
parameters of the cooling schedule. See \code{\link[spsann]{scheduleSPSANN}}.}

\item{plotit}{Logical for plotting the optimization results, including
a) the progress of the objective function, and b) the starting (gray) and
current system configuration (black), and the maximum jitter in the 
x- and y-coordinates. The plots are updated at each 10 jitters. Defaults to
\code{plotit = FALSE}.}

\item{track}{Logical value. Should the evolution of the energy state be 
recorded and returned with the result? If \code{track = FALSE} (the default),
only the starting and ending energy states are returned with the results.}

\item{boundary}{SpatialPolygon defining the boundary of the spatial domain. 
If missing and \code{plotit = TRUE}, \code{boundary} is estimated from
\code{candi}.}

\item{progress}{Type of progress bar that should be used, with options 
\code{"txt"}, for a text progress bar in the R console, \code{"tk"}, to put
up a Tk progress bar widget, and \code{NULL} to omit the progress bar.
A Tk progress bar widget is usefull when using parallel processors. Defaults
to \code{progress = "txt"}.}

\item{verbose}{Logical for printing messages about the progress of the
optimization. Defaults to \code{verbose = FALSE}.}
}
\value{
\code{optimDIST} returns an object of class \code{OptimizedSampleConfiguration}: the optimized sample
configuration with details about the optimization.
 
\code{objDIST} returns a numeric value: the energy state of the sample configuration -- the objective
function value.
}
\description{
Optimize a sample configuration for spatial trend identification and  estimation. A criterion is defined 
so that the sample reproduces the marginal distribution of the covariates (\bold{DIST}).
}
\note{
The distance between two points is computed as the Euclidean distance between
them. This computation assumes that the optimization is operating in the 
two-dimensional Euclidean space, i.e. the coordinates of the sample points 
and candidate locations should not be provided as latitude/longitude. 
The \pkg{spsann}-package has no mechanism to check if the coordinates are
projected: the user is responsible for making sure that this requirement
is attained.
}
\section{Marginal distribution of covariates}{


\emph{Equal-area} marginal sampling strata are defined using the unique
sample quantiles estimated with \code{\link[stats]{quantile}} using a 
discontinuous function (\code{type = 3}). This is to avoid creating empty
marginal sampling strata. 

\emph{Equal-range} marginal sampling strata are defined by breaking the range
of covariate values into pieces of equal size. Any empty marginal sampling
strata is merged with the closest non-empty marginal sampling strata.
}
\examples{
require(sp)
data(meuse.grid)
candi <- meuse.grid[, 1:2]
covars <- meuse.grid[, 5]
schedule <- scheduleSPSANN(initial.temperature = 1, chains = 1,
                           x.max = 1540, y.max = 2060, x.min = 0, 
                           y.min = 0, cellsize = 40)
set.seed(2001)
res <- optimDIST(points = 10, candi = candi, covars = covars,
                 use.coords = TRUE, schedule = schedule)
objSPSANN(res) -
  objDIST(points = res, candi = candi, covars = covars, use.coords = TRUE)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\references{
Edzer Pebesma, Jon Skoien with contributions from Olivier Baume, A. Chorti, 
D.T. Hristopulos, S.J. Melles and G. Spiliopoulos (2013). 
\emph{intamapInteractive: procedures for automated interpolation - methods 
only to be used interactively, not included in \code{intamap} package.} R 
package version 1.1-10.

van Groenigen, J.-W. \emph{Constrained optimization of spatial sampling: 
a geostatistical approach.} Wageningen: Wageningen University, p. 148, 1999.

Walvoort, D. J. J.; Brus, D. J. & de Gruijter, J. J. An R package for 
spatial coverage sampling and random sampling from compact geographical 
strata by k-means. \emph{Computers & Geosciences}. v. 36, p. 1261-1267, 2010.

Cramér, H. \emph{Mathematical methods of statistics}. Princeton: Princeton University Press, p. 575, 1946.

Everitt, B. S. \emph{The Cambridge dictionary of statistics}. Cambridge: Cambridge University Press, p. 432,
2006.

Hyndman, R. J.; Fan, Y. Sample quantiles in statistical packages. \emph{The American Statistician}, v. 50, 
p. 361-365, 1996.

Minasny, B.; McBratney, A. B. A conditioned Latin hypercube method for sampling in the presence of 
ancillary information. \emph{Computers & Geosciences}, v. 32, p. 1378-1388, 2006.

Minasny, B.; McBratney, A. B. Conditioned Latin Hypercube Sampling for calibrating soil sensor data to soil
properties. Chapter 9. Viscarra Rossel, R. A.; McBratney, A. B.; Minasny, B. (Eds.) \emph{Proximal Soil
Sensing}. Amsterdam: Springer, p. 111-119, 2010.

Roudier, P.; Beaudette, D.; Hewitt, A. A conditioned Latin hypercube sampling algorithm incorporating
operational constraints. \emph{5th Global Workshop on Digital Soil Mapping}. Sydney, p. 227-231, 2012.
}
\seealso{
\code{\link[clhs]{clhs}}, \code{\link[spsann]{optimACDC}}
}
\concept{
simulated annealing

spatial trend
}
\keyword{iteration}
\keyword{optimize}
\keyword{spatial}

