% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{OptimizedSampleConfiguration-class}
\alias{OptimizedSampleConfiguration}
\alias{OptimizedSampleConfiguration-class}
\title{An S4 class}
\description{
An S4 class
}
\section{Slots}{

\describe{
\item{\code{points}}{}

\item{\code{spsann}}{}

\item{\code{objective}}{}
}}

