% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optimPPL.R
\name{optimPPL}
\alias{objPairs}
\alias{objPoints}
\alias{optimPPL}
\alias{pairsPerLag}
\alias{pointsPerLag}
\title{Optimization of sample patterns for for variogram estimation}
\usage{
optimPPL(points, candidates, lags = 7, lags.type = "exponential",
  lags.base = 2, cutoff = NULL, criterion = "distribution",
  pre.distri = NULL, x.max, x.min, y.max, y.min, iterations = 10000,
  acceptance = list(initial = 0.99, cooling = iterations/10),
  stopping = list(max.count = iterations/10), plotit = TRUE, boundary,
  progress = TRUE, verbose = TRUE)

objPoints(points, candidates, lags = 7, lags.type = "exponential",
  lags.base = 2, cutoff = NULL, criterion = "distribution",
  pre.distri = NULL)

pointsPerLag(points, candidates, lags = 7, lags.type = "exponential",
  lags.base = 2, cutoff = NULL)
}
\arguments{
\item{points}{Integer. The number of points. Alternatively, a data.frame or
matrix with three columns: 1) the identification of each point
(1, 2, ..., n), 2) the x coordinates, and 3) the y coordinates. The
coordinates must be projected. If a data.frame or matrix is used,
\code{points} must be a subset of \code{candidates}.}

\item{candidates}{data.frame or matrix. The candidate locations for the
sample points.}

\item{lags}{Integer. The number of lag distance classes. Alternatively, a
vector of numeric values with the lower and upper limits of each lag
distance class. The lowest value must be larger than zero, e.g. 0.0001.
Defaults to \code{lags = 7}.}

\item{lags.type}{Character. The type of lag distance classes. Available
options are \code{"equidistant"} and \code{"exponential"}. Defaults to
\code{lags.type = "exponential"}. See \sQuote{Details} for more information.}

\item{lags.base}{Numeric. Base of the exponential expression used to
create the exponential lag distance classes. Defaults to
\code{lags.base = 2}. See \sQuote{Details} for more information.}

\item{cutoff}{Numeric. The maximum distance up to which lag distance classes
are created. Used only when \code{lags} is an integer. See \sQuote{Details}
for more information.}

\item{criterion}{Character. The measure to be used to describe the energy
state of the current system configuration. Available options are
\code{"minimum"} and \code{"distribution"}. Defaults to
\code{objective = "minimum"}. See \sQuote{Details} for more information.}

\item{pre.distri}{Numeric vector. The pre-specified distribution of points
or point-pair with which the observed counts of points or point-pairs per
lag distance class is compared. Used only when
\code{criterion = "distribution"}. Defaults to a uniform distribution. See
\sQuote{Details} for more information.}

\item{x.max,x.min,y.max,y.min}{Numeric. The minimum and maximum quantity of
random noise to be added to the x and y coordinates. The minimum quantity
must be equal to, at least, the minimum distance between two neighboring
candidate locations. The units are the same as of the coordinates.}

\item{iterations}{Integer. The maximum number of iterations that should be
used for the optimization.}

\item{acceptance}{List with two sub-arguments: \code{initial} and
\code{cooling}. \code{initial} is a numeric value between 0 and 1 defining
the initial acceptance probability. Defaults to \code{initial = 0.99}.
\code{cooling} is a numeric value defining the exponential factor by with
the acceptance probability decreases at each iteration. Defaults to
\code{cooling = iterations / 10}.}

\item{stopping}{List with one sub-argument: \code{max.count}.
\code{max.count} is an integer value defining the maximum allowable number
of iterations without improvement of the objective function value. Defaults
to \code{max.count = iterations / 10}.}

\item{plotit}{Logical for ploting the optimization results. This includes
a) the progress of the objective function values and acceptance
probabilities, and b) the original points, the perturbed points and the
progress of the maximum perturbation in the x and y coordinates. The plots
are updated at each 10 iterations. Defaults to \code{plotit = TRUE}.}

\item{boundary}{SpatialPolygon. The boundary of the spatial domain.
Mandatory if \code{plotit = TRUE}.}

\item{progress}{Logical for printing a progress bar. Defaults to
\code{progress = TRUE}.}

\item{verbose}{Logical for printing messages about the progress of the
optimization.}
}
\value{
\code{optimPPL} returns a matrix: the optimized sample pattern with
the evolution of the energy state during the optimization as an attribute.

\code{pointsPerLag} and \code{pairsPerLag} return a data.frame with three
columns: a) the lower and b) upper limits of each lag, and c) the number of
points or point-pairs per lag.

\code{objPoints} and \code{objPairs} return a numeric value depending on the
choice of \code{criterion}. If \code{criterion = "distribution"}, the sum of
the differences between the pre-specified and observed distribution of counts
of points or point-pairs per lag. If \code{criterion = "minimum"}, the
inverse of the minimum count of points or point pairs over all lags
multiplied by a constant.
}
\description{
Optimize a sample pattern for variogram estimation. The criterion used is
the number of points or point-pairs per lag distance class (\code{optimPPL}).
\code{pointsPerLag} and \code{pairsPerLag} count the number of points or
point-pairs per lag distance class. \code{objPoints} and \code{objPairs}
compute the deviation of the observed distribution of counts from a
pre-specified distribution, or the minimum number of points or point pairs
observed over all lag distance classes.
}
\details{
\subsection{Distances}{
Euclidean distances between points are calculated. This computation requires
the coordinates to be projected. The user is responsible for making sure that
this requirement is attained.
}
\subsection{Distribution}{
Using the default uniform distribution means that the number of
\strong{point-pairs} per lag is equal to
\eqn{n \times (n - 1) / (2 \times lag)}, where \eqn{n} is the total number
of points and \eqn{lag} is the number of lags.

Using the default uniform distribution means that the number of
\strong{points} per lag is equal to the total number of points. This is the
same as expecting that each point contributes to every lag.

Distributions other than the available options can be easily implemented
changing the arguments \code{lags}, \code{lags.base} and \code{pre.distri}.
}
\subsection{Type of lags}{
Two types of lags can be created by default. The first are evenly spaced
lags (\code{lags.type = "equidistant"}). They are created by simply dividing
the distance interval from 0.0001 to \code{cutoff} by the required number of
lags. The minimum value of 0.0001 guarantees that a point does not form a
pair with itself.

The second type of lags is defined by exponential spacings
(\code{lags.type = "exponential"}). The spacings are defined by the base
\eqn{b} of the exponential expression \eqn{b^n}, where \eqn{n} is the
required number of lags. The base is defined using the argument
\code{lags.base}.
}
\subsection{Criteria}{
There are two optimizing criteria implemented. The first is called using
\code{criterion = "distribution"} and is used to minimize the sum of
differences between a pre-specified distribution and the observed
distribution of points or point-pairs per lag.

The second criterion is called using \code{criterion = "minimum"}. It
corresponds to maximizing the minimum number of points or point-pairs
observed over all lags.
}
}
\examples{
require(sp)
data(meuse)
meuse <- as.matrix(meuse[, 1:2])
meuse <- matrix(cbind(c(1:dim(meuse)[1]), meuse), ncol = 3)
pointsPerLag(meuse, cutoff = 1000)
objPoints(meuse, cutoff = 1000)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\references{
Bresler, E.; Green, R. E. \emph{Soil parameters and sampling scheme for
characterizing soil hydraulic properties of a watershed}. Honolulu:
University of Hawaii at Manoa, p. 42, 1982.

Marler, R. T.; Arora, J. S. Function-transformation methods for
multi-objective optimization. \emph{Engineering Optimization}. v. 37, p.
551-570, 2005.

Russo, D. Design of an optimal sampling network for estimating the variogram.
\emph{Soil Science Society of America Journal}. v. 48, p. 708-716, 1984.

Truong, P. N.; Heuvelink, G. B. M.; Gosling, J. P. Web-based tool for expert
elicitation of the variogram. \emph{Computers and Geosciences}. v. 51, p.
390-399, 2013.

Warrick, A. W.; Myers, D. E. Optimization of sampling locations for variogram
calculations. \emph{Water Resources Research}. v. 23, p. 496-500, 1987.
}
\concept{
simulated annealing
}
\keyword{optimize}
\keyword{spatial}

