% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osc-class.R
\docType{class}
\name{OptimizedSampleConfiguration-class}
\alias{OptimizedSampleConfiguration}
\alias{OptimizedSampleConfiguration-class}
\title{Class \sQuote{OptimizedSampleConfiguration}}
\description{
An S4 class
}
\section{Slots}{

\describe{
\item{\code{points}}{}

\item{\code{spsann}}{}

\item{\code{objective}}{}
}}
\examples{
osc <- new("OptimizedSampleConfiguration")
getSlots("OptimizedSampleConfiguration")
is(osc, "OptimizedSampleConfiguration")
}

