% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scheduleSPSANN.R
\name{scheduleSPSANN}
\alias{scheduleSPSANN}
\title{\pkg{spsann} cooling schedule}
\usage{
scheduleSPSANN(initial.acceptance = 0.95, initial.temperature = 0.001,
  temperature.decrease = 0.95, chains = 500, chain.length = 1,
  stopping = 10, x.max, x.min, y.max, y.min)
}
\arguments{
\item{initial.acceptance}{Numeric value between 0 and 1 defining the initial
acceptance probability. Defaults to \code{initial.acceptance = 0.95}.}

\item{initial.temperature}{Numeric value larger than 0 defining the initial
temperature of the system. Defaults to \code{initial.temperature = 0.001}.}

\item{temperature.decrease}{Numeric value between 0 and 1 defining the factor
by which the temperature is decreased at the end of each Markov chain.
Defaults to \code{temperature.decrease = 0.95}.}

\item{chains}{Integer value defining the maximum number of Markov chains.
Defaults to \code{chains = 500}.}

\item{chain.length}{Integer value defining the length of each Markov chain
relative to the number of points. Defaults to \code{chain.length = 1}, i.e.
one time the number of points.}

\item{stopping}{Integer value defining the maximum allowable number of
Markov chains without improvement of the objective function value. Defaults
to \code{stopping = 10}.}

\item{x.max}{Numeric value. The minimum and maximum
quantity of random noise to be added to the projected x- and y-coordinates.
The minimum quantity should be equal to, at least, the minimum distance
between two neighbouring candidate locations. The units are the same as of
the projected x- and y-coordinates. If missing, they are estimated from
\code{candi}.}

\item{x.min}{Numeric value. The minimum and maximum
quantity of random noise to be added to the projected x- and y-coordinates.
The minimum quantity should be equal to, at least, the minimum distance
between two neighbouring candidate locations. The units are the same as of
the projected x- and y-coordinates. If missing, they are estimated from
\code{candi}.}

\item{y.max}{Numeric value. The minimum and maximum
quantity of random noise to be added to the projected x- and y-coordinates.
The minimum quantity should be equal to, at least, the minimum distance
between two neighbouring candidate locations. The units are the same as of
the projected x- and y-coordinates. If missing, they are estimated from
\code{candi}.}

\item{y.min}{Numeric value. The minimum and maximum
quantity of random noise to be added to the projected x- and y-coordinates.
The minimum quantity should be equal to, at least, the minimum distance
between two neighbouring candidate locations. The units are the same as of
the projected x- and y-coordinates. If missing, they are estimated from
\code{candi}.}
}
\value{
A list with a set of control parameters of the cooling schedule.
}
\description{
Set the control parameters for the cooling schedule of \pkg{spsann}
functions.
}
\examples{
schedule <- scheduleSPSANN()
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\references{
Aarts, E. H. L. & Korst, J. H. M. Boltzmann machines for travelling salesman
problems. \emph{European Journal of Operational Research}. v. 39, p. 79-95,
1989.
}
\seealso{
\code{\link[spsann]{optimACDC}}, \code{\link[spsann]{optimCORR}},
\code{\link[spsann]{optimDIST}}, \code{\link[spsann]{optimMKV}},
\code{\link[spsann]{optimMSSD}}, \code{\link[spsann]{optimPPL}},
\code{\link[spsann]{optimSPAN}}, \code{\link[spsann]{optimUSER}}.
}

