% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spJitterFinite}
\alias{spJitterFinite}
\title{Random perturbation of spatial points}
\usage{
spJitterFinite(points, candidates, x.max, x.min, y.max, y.min, which.point)
}
\arguments{
\item{points}{Integer. The number of points. Alternatively, a data.frame or
matrix with three columns: 1) the identification of each point
(1, 2, ..., n), 2) the x coordinates, and 3) the y coordinates. The
coordinates must be projected. If a data.frame or matrix is used,
\code{points} must be a subset of \code{candidates}.}

\item{candidates}{data.frame or matrix. The candidate locations for the
sample points.}

\item{x.max,x.min,y.max,y.min}{Numeric. The minimum and maximum quantity of
random noise to be added to the x and y coordinates. The minimum quantity
must be equal to, at least, the minimum distance between two neighboring
candidate locations. The units are the same as of the coordinates.}

\item{which.point}{Integer values defining which point should be perturbed.
The current version accepts only one point to be perturbed at a time. See
\sQuote{Details} for more information.}
}
\value{
A matrix with the jittered coordinates of the points.
}
\description{
Perturbate the coordinates of spatial points (\sQuote{jittering}).
}
\details{
There are two ways of jittering the coordinates. They differ on how the the
set of candidate locations is defined. The first method uses a finite set of
candidate locations for the perturbed points. The current version does not
accept using an infinite set of candidate locations.
}
\note{
The current version does not accept using an infinite set of candidate
locations or to perturb more than one point at a time.
}
\examples{
require(sp)
data(meuse.grid)
meuse.grid <- as.matrix(meuse.grid[, 1:2])
meuse.grid <- matrix(cbind(1:dim(meuse.grid)[1], meuse.grid), ncol = 3)
pts1 <- sample(c(1:dim(meuse.grid)[1]), 155)
pts2 <- meuse.grid[pts1, ]
pts3 <- spJitterFinite(points = pts2, candidates = meuse.grid, x.min = 40,
                      x.max = 100, y.min = 40, y.max = 100, which.point = 10)
plot(meuse.grid[, 2:3], asp = 1, pch = 15, col = "gray")
points(pts2[, 2:3], col = "red", cex = 0.5)
points(pts3[, 2:3], pch = 19, col = "blue", cex = 0.5)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\references{
Edzer Pebesma, Jon Skoien with contributions from Olivier Baume, A. Chorti,
D.T. Hristopulos, S.J. Melles and G. Spiliopoulos (2013).
\emph{intamapInteractive: procedures for automated interpolation - methods
only to be used interactively, not included in intamap package.} R package
version 1.1-10. \url{http://CRAN.R-project.org/package=intamapInteractive}

van Groenigen, J.-W. \emph{Constrained optimization of spatial sampling:
a geostatistical approach.} Wageningen: Wageningen University, p. 148, 1999.
}
\seealso{
\code{ssaOptim}, \code{\link[sp]{zerodist}},
\code{\link[base]{jitter}}, \code{\link[geoR]{jitter2d}}.
}
\concept{
jitter perturb
}
\keyword{iteration}
\keyword{spatial}

