% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SPSANNtools.R
\name{objSPSANN}
\alias{SPSANNtools}
\alias{objSPSANN}
\title{Auxiliary tools}
\usage{
objSPSANN(osc, at = "end", n = 1)
}
\arguments{
\item{osc}{Optimized Sample Configuration.}

\item{at}{Point of the optimization at which the energy state should be
returned. Available options: \code{"start"}, for the start, and \code{"end"},
for the end of the optimization. Defaults to \code{at = "end"}.}

\item{n}{Number of instances that should be returned. Defaults to
\code{n = 1}.}
}
\description{
Auxiliary tools used in the optimization of sample configurations using
spatial simulated annealing.
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}

