% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minmaxPareto.R
\name{minmaxPareto}
\alias{minmaxPareto}
\title{Pareto maximum and minimum values}
\usage{
minmaxPareto(osc, candi, covars)
}
\arguments{
\item{osc}{A list with the optimized sample configurations (OSC). Each OSC
must be named after the objective function with which it has been optimized.
For example, \code{osc = list(CORR = osc_corr, DIST = osc_dist)}.}

\item{candi}{Data frame or matrix with the candidate locations for the
jittered points. \code{candi} must have two columns in the following
order: \code{[, "x"]} the projected x-coordinates, and \code{[, "y"]} the
projected y-coordinates.}

\item{covars}{Data frame or matrix with the covariates in the columns.}

\item{...}{Other arguments required to compute the objective function value.
See the help pages of the respective objective functions to see which
arguments are needed.}
}
\value{
A data frame with the Pareto maximum and minimum values.
}
\description{
Compute the Pareto maximum and minimum values of the objective functions
that compose a multi-objective optimization problem (MOOP). The Pareto
maximum and minimum values are then used to transform the objective
functions using the \emph{upper-lower-bound approach} so that they can be
aggregated into a single \emph{utility function} using the \emph{weighted
sum method}.
}
\examples{
\dontrun{
# This example takes more than 5 seconds
require(sp)
data(meuse.grid)
candi <- meuse.grid[, 1:2]
covars <- meuse.grid[, c(1, 2)]

# CORR
schedule <- scheduleSPSANN(initial.acceptance = 0.1, chains = 1,
                           x.max = 1540, y.max = 2060, x.min = 0,
                           y.min = 0, cellsize = 40)
set.seed(2001)
osc_corr <- optimCORR(points = 10, candi = candi, covars = covars,
                      schedule = schedule)

# DIST
set.seed(2001)
osc_dist <- optimDIST(points = 10, candi = candi, covars = covars,
                      schedule = schedule)

# PPL
set.seed(2001)
osc_ppl <- optimPPL(points = 10, candi = candi, schedule = schedule)

# MSSD
set.seed(2001)
osc_mssd <- optimMSSD(points = 10, candi = candi, schedule = schedule)

# Pareto
pareto <- minmaxPareto(osc = list(DIST = osc_dist, CORR = osc_corr,
                                  PPL = osc_ppl, MSSD = osc_mssd),
                       candi = candi, covars = covars)
pareto
}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\seealso{
\code{\link[spsann]{optimACDC}}, \code{SPAN}
}

