% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optimMSSD.R
\name{optimMSSD}
\alias{objMSSD}
\alias{optimMSSD}
\title{Optimization of sample patterns for spatial interpolation}
\usage{
optimMSSD(points, candidates, x.max, x.min, y.max, y.min, iterations = 10000,
  acceptance = list(initial = 0.99, cooling = iterations/10),
  stopping = list(max.count = iterations/10), plotit = TRUE, boundary,
  progress = TRUE, verbose = TRUE)

objMSSD(candidates, points)
}
\arguments{
\item{points}{Integer. The number of points. Alternatively, a data.frame or
matrix with three columns: 1) the identification of each point
(1, 2, ..., n), 2) the x coordinates, and 3) the y coordinates. The
coordinates must be projected. If a data.frame or matrix is used,
\code{points} must be a subset of \code{candidates}.}

\item{candidates}{data.frame or matrix. The candidate locations for the
sample points.}

\item{x.max,x.min,y.max,y.min}{Numeric. The minimum and maximum quantity of
random noise to be added to the x and y coordinates. The minimum quantity
must be equal to, at least, the minimum distance between two neighboring
candidate locations. The units are the same as of the coordinates.}

\item{iterations}{Integer. The maximum number of iterations that should be
used for the optimization.}

\item{acceptance}{List with two sub-arguments: \code{initial} and
\code{cooling}. \code{initial} is a numeric value between 0 and 1 defining
the initial acceptance probability. Defaults to \code{initial = 0.99}.
\code{cooling} is a numeric value defining the exponential factor by with
the acceptance probability decreases at each iteration. Defaults to
\code{cooling = iterations / 10}.}

\item{stopping}{List with one sub-argument: \code{max.count}.
\code{max.count} is an integer value defining the maximum allowable number
of iterations without improvement of the objective function value. Defaults
to \code{max.count = iterations / 10}.}

\item{plotit}{Logical for ploting the optimization results. This includes
a) the progress of the objective function values and acceptance
probabilities, and b) the original points, the perturbed points and the
progress of the maximum perturbation in the x and y coordinates. The plots
are updated at each 10 iterations. Defaults to \code{plotit = TRUE}.}

\item{boundary}{SpatialPolygon. The boundary of the spatial domain.
Mandatory if \code{plotit = TRUE}.}

\item{progress}{Logical for printing a progress bar. Defaults to
\code{progress = TRUE}.}

\item{verbose}{Logical for printing messages about the progress of the
optimization.}
}
\value{
\code{objMSSD} returns a numeric value: the MSSD between a set of points and
all grid cells.

\code{optimMSSD} returns a matrix: the optimized sample pattern with
the evolution of the energy state during the optimization as an attribute.
}
\description{
Optimize a sample pattern for spatial interpolation. The criterion used is
the mean squared shortest distance (\code{optimMSSD}). \code{objMSSD}
computes the MSSD between a set of points and all grid cells.
}
\details{
Euclidean distances between points are calculated. This computation requires
the coordinates to be projected. The user is responsible for making sure that
this requirement is attained.
}
\examples{
require(sp)
data(meuse.grid)
meuse.grid <- as.matrix(meuse.grid[, 1:2])
meuse.grid <- matrix(cbind(c(1:dim(meuse.grid)[1]), meuse.grid), ncol = 3)
points <- sample(c(1:dim(meuse.grid)[1]), 155)
points <- meuse.grid[points, ]
objMSSD(meuse.grid, points)
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
\references{
Brus, D. J.; de Gruijter, J. J.; van Groenigen, J. W. Designing spatial
coverage samples using the k-means clustering algorithm. In: P. Lagacherie,
A. M.; Voltz, M. (Eds.) \emph{Digital soil mapping - an introductory
perspective}. Elsevier, v. 31, p. 183-192, 2006.

de Gruijter, J. J.; Brus, D.; Bierkens, M.; Knotters, M. \emph{Sampling for
natural resource monitoring}. Berlin: Springer, p. 332, 2006.

Walvoort, D. J. J.; Brus, D. J.; de Gruijter, J. J. An R package for spatial
coverage sampling and random sampling from compact geographical strata by
k-means. \emph{Computers and Geosciences}. v. 36, p. 1261-1267, 2010.
}
\seealso{
\code{\link[raster]{distanceFromPoints}}, \code{\link[spcosa]{stratify}}.
}
\concept{
simulated annealing
}
\keyword{optimize}
\keyword{spatial}

